refactor(ui): categorize ui directory
