+++ packages/ui/src/components/index.ts
@@ -1,0 +1,3 @@
+export { Button } from './button';
+export { Paragraph } from './paragraph';
+export type { ParagraphProps } from './paragraph';
