+++ packages/ui/src/index.ts
@@ -1,8 +1,9 @@
 'use client';
-// NOTE: use client 명시를 해주지 않으면 서버 컴포넌트에서 클라이언트 컴포넌트를 import 할 수 없다는 에러 발생
-// 버그에 가까운 동작
 
-export * from './button';
-export * from './paragraph';
+export { Button, Paragraph } from './components';
+export type { ParagraphProps } from './components';
+
+export { colors, colorsMap, typography } from './systems';
+export type { Color, Typography } from './systems';
+
 export type * from './types';
-export * from './colors';
