+++ packages/ui/src/systems/index.ts
@@ -1,0 +1,4 @@
+export { typography } from './typography';
+export type { Typography } from './typography';
+export { colors, colorsMap } from './colors';
+export type { Color } from './colors';
